add minimum viable bottom
It is a series of pillars attached to a base, like the mockups but
with screw hole pillars instead of sa-keycaps. I left the actual screw
hole pillars off of the top, so that they can be printed in their
native orientation, because the direction of the layer lines matters
a good deal when melting threaded inserts in. They will need to be
glued on.

This is the simplest shape the frame can be screwed to, discarding all
consideration of appearance or elegant function, because my thumbs
hurt and I need a keyboard.