screw hole pillar fixes (closes #64)
This change fixes the location of the screw hole pillars of the
mvbottom, placement of which was not taking into account the z-munging
that the frame screw holes had, with which the screw hole pillars are
intended to mate. Screw hole pillars (plus a tolerance) are now
differenced from screw holes (#64). An annoying circular dependency is
circumvented by an annoying manipulation.